pyproject.toml
[project]
name = "rick-and-morty-api-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.6"
dependencies = [
    "allure-pytest>=2.15.0",
    "pytest>=8.4.2",
    "pytest-html>=4.1.1",
    "requests>=2.32.5",
    "pytest-playwright>=0.7.0",
    "pytest-bdd>=8.1.0",
]
